% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-definitions.R
\docType{data}
\name{ego}
\alias{ego}
\title{Facebook social circles}
\format{An igraph package undirected graph object with 4039 vertices and 88234 edges. The
graph includes a force-directed layout with vertices colored by the \code{\link{cluster_fast_greedy}}
algorithm from the igraph package.}
\source{
Stanford SNAP network repository
\url{http://snap.stanford.edu/data/facebook_combined.txt.gz}
}
\usage{
data(ego)
}
\description{
A facebook social network subgraph obtained from the Stanford SNAP repository.
}
\references{
J. McAuley and J. Leskovec. Learning to Discover Social Circles in Ego Networks. NIPS, 2012.
}
\keyword{datasets}

