% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gcol}
\alias{gcol}
\title{A basic internal color format parser}
\usage{
gcol(x)
}
\arguments{
\item{x}{a character-valued color name}
}
\value{
a list of 3-hex-digit color values and scalar numeric alpha values
}
\description{
A basic internal color format parser
}

