% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph2Matrix}
\alias{graph2Matrix}
\title{Convert from node and edge graph representation to a sparse adjacency matrix representation}
\usage{
graph2Matrix(nodes, edges, symmetric = TRUE)
}
\arguments{
\item{nodes}{A data frame with at least a column named "id" as used by \code{\link{graphjs}}. The size of the matrix is determined by  number of rows in the data frame.}

\item{edges}{A data frame with at least the columns "from" and "to" referring to edges between ids in the \code{nodes} data frame. If the data frame includes a numeric "size" variable then graph is assumed to be weighted and the corresponding matrix entries are set to the size values.}

\item{symmetric}{Set to \code{FALSE} for directed graphs, or leave as \code{TRUE} for undirected graphs.}
}
\value{
A sparse matrix
}
\description{
Convert from node and edge graph representation to a sparse adjacency matrix representation
}
\examples{
data(LeMis)
M <- graph2Matrix(LeMis$nodes, LeMis$edges)
G <- matrix2graph(M)
}
\seealso{
\code{\link{graphjs}}, \code{\link{graph2Matrix}}
}

