% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hello}
\alias{hello}
\title{hello}
\usage{
hello(height = NULL, width = NULL)
}
\arguments{
\item{height}{The container div height.}

\item{width}{The container div width.}
}
\description{
Threejs hello world example
}
\examples{
## dontrun
# A stand-alone example
library("threejs")
hello()
}
\references{
The threejs project \url{http://threejs.org}.
}
\seealso{
scatterplot3d, rgl, scatterplotx3
}

