% Generated by roxygen2 (4.0.2): do not edit by hand
\name{helloThree.js}
\alias{helloThree.js}
\title{helloThree.js  Three.js hello world example.}
\usage{
helloThree.js(img, ambient = "#999", specular = "#ffa", shininess = 10,
  height = NULL, width = NULL)
}
\arguments{
\item{img}{Either a character string indicating a path to a file
(shiny apps), or an image texture prepared by \code{texture}.}

\item{ambient}{The intensity/color of object light scattering.}

\item{specular}{The intensity/color of object specular reflection.}

\item{shininess}{The shininess of the object.}

\item{height}{The container div height.}

\item{width}{The container div width.}
}
\description{
Three.js hello world example, maps an image onto a rotating sphere.
}
\examples{
## dontrun
# Stand-alone examples:
library("threejs")
helloThree.js(texture(system.file("img/Rlogo.jpg",package="threejs")))
helloThree.js(texture(system.file("images/moon.jpg",package="threejs")))

# A shiny example:
library("shiny")
runApp(system.file("examples/hello",package="threejs"), launch.browser=FALSE)
}
\references{
The threejs project \url{http://threejs.org}.
The corresponding javascript file in
\code{ system.file("htmlwidgets/hello.js",package="threejs")}.
}

