% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{light_ambient}
\alias{light_ambient}
\title{Plot illumination}
\usage{
light_ambient(color = "#eeeeee")
}
\arguments{
\item{color}{the ambient light color}
}
\value{
An object for use with the \code{lights} argument in \code{scatterplot3js} and \code{graphjs}.
}
\description{
Plot illumination
}
