% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{light_directional}
\alias{light_directional}
\title{Plot illumination}
\usage{
light_directional(color = "#eeeeee", position = c(0, 0, 0))
}
\arguments{
\item{color}{the light color}

\item{position}{the light position as an (x, y, z) coordinate vector with entries in [-1, 1]}
}
\value{
An object for use with the \code{lights} argument in \code{scatterplot3js} and \code{graphjs}.
}
\description{
Plot illumination
}
