% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{points3d}
\alias{points3d}
\title{Add points to a 3D scatterplot}
\usage{
points3d(s, ...)
}
\arguments{
\item{s}{A scatterplot object returned by \code{\link{scatterplot3js}}.}

\item{...}{Optional new point arguments (see examples) following \code{\link{scatterplot3js}} syntax.}
}
\value{
A new scatterplot htmlwidget object.
}
\description{
Add points to a 3D scatterplot
}
\note{
This function replaces the old \code{points3d} approach used by \code{scatterplot3d}.
}
\examples{
\dontrun{
 x <- rnorm(5)
 y <- rnorm(5)
 z <- rnorm(5)
 scatterplot3js(x, y, z, pch="o") \%>\%
   points3d(x + 0.2, y + 0.2, z, color="red", pch=paste("point", 1:5))
}
}

