% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scatterplotThree}
\alias{scatterplotThree}
\title{scatterplotThree Threejs 3D scatterplot example}
\usage{
scatterplotThree(x, height = NULL, width = NULL, axis = TRUE,
  tick.marks = TRUE, num.ticks = c(6, 6, 6), color = "steelblue",
  size = 0.15, grid = TRUE)
}
\arguments{
\item{x}{A data matrix with three columns corresponding to the x,y,z
coordinate axes. Column labels, if present, are used as axis labels.}

\item{height}{The container div height.}

\item{width}{The container div width.}

\item{tick.marks}{Set FALSE to disable display of tick marks.}

\item{num.ticks}{A three-element vector with the suggested number of
ticks to display per axis.}

\item{color}{Either a single valid RGB, hex or named color name, or
a vector of valid color names of length \code{nrow(x)}.}

\item{size}{The plot point radius, either as a single number or a
vector of sizes of length \code{nrow(x)}.}

\item{grid}{Set FALSE to disable display of a grid.}
}
\description{
A basic 3D scatterplot using d3.js and three.js.
}
\examples{
## dontrun
# A stand-alone example
set.seed(1)
x <- matrix(rnorm(100*3),ncol=3)
scatterplotThree(x)

# A shiny example
runApp(system.file("examples/shiny",package="three"))
}
\references{
The threejs project \url{http://threejs.org}.
}
\seealso{
scatterplot3d, rgl, scatterplotx3
}

