% Generated by roxygen2 (4.0.2): do not edit by hand
\name{texture}
\alias{texture}
\title{texture Convert a matrix or raster image to a three.js texture}
\usage{
texture(data, alpha = 1)
}
\arguments{
\item{data}{A data matrix representing a grayscale image or a
raster image object with optional alpha channel.}

\item{alpha}{The default alpha level in [0,1] to use if not
specified in the raster image format.}
}
\description{
Convert matrix or raster image image representations in R into
JSON-formatted arrays suitable for use as three.js DataTexture
RGBA textures.
}
\examples{
## dontrun
library("png")
img <- readPNG(system.file("img", "Rlogo.png", package="png"))
texture(img)
}
\references{
The threejs project \url{http://threejs.org}.
}

