% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/globe.R, R/scatterplot.R, R/threejs.R
\name{globeOutput}
\alias{globeOutput}
\alias{renderGlobe}
\alias{renderScatterplotThree}
\alias{scatterplotThreeOutput}
\alias{threejs-shiny}
\title{Shiny bindings for threejs widgets}
\usage{
globeOutput(outputId, width = "100\%", height = "600px")

renderGlobe(expr, env = parent.frame(), quoted = FALSE)

scatterplotThreeOutput(outputId, width = "100\%", height = "500px")

renderScatterplotThree(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width,height}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{expr}{An expression that generates threejs graphics.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
  is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using threejs widgets within Shiny
applications and interactive Rmd documents.
}

