% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\docType{methods}
\name{vertices,scatterplotThree-method}
\alias{vertices,scatterplotThree-method}
\title{Extract a matrix of vertex coordinates from a threejs widget}
\usage{
\S4method{vertices}{scatterplotThree}(...)
}
\arguments{
\item{...}{a \code{scatterplotThree} object from the threejs package.}
}
\description{
Extract a matrix of vertex coordinates from a threejs widget
}
\seealso{
points3d
}

