% Generated by roxygen2 (4.0.2): do not edit by hand
\name{globe.js}
\alias{globe.js}
\title{globe.js Three.js globe example.}
\usage{
globe.js(img, lat, long, color = "red", value = 40, bodycolor = "#0000ff",
  emissive = "#0000ff", lightcolor = "#9999ff", height = NULL,
  width = NULL)
}
\arguments{
\item{img}{A character string representing an image file path, or a dataURI
image prepared by the \code{texture} function, of an image to plot on the globe.}

\item{lat}{Data point latitudes (negative values indicate south, positive north).}

\item{long}{Data point longitudes, must be of same length as \code{lat} (negative values indicate west, positive east).}

\item{color}{Either a single color value indicating the color of all data points, or a vector of values of length x.lat indicating color of each point.}

\item{value}{Either a single value indicating the height of all data points, or a vector of values of length x.lat indicating height of each point.}

\item{bodycolor}{The diffuse reflective color of the globe object.}

\item{emissive}{The emissive color of the globe object.}

\item{lightcolor}{The color of the ambient light in the scene.}

\item{height}{The container div height.}

\item{width}{The container div width.}
}
\description{
Three.js example that maps points and an image on a globe.
}
\note{
The \code{img} argument may be a relative file location string pointing to
an image file in shiny applications (see the examples below). For non-shiny
use, prepare the image file as a dataURI with the \code{texture} function
(see the examples).
}
\examples{
## dontrun
# A shiny example:
library("shiny")
runApp(system.file("examples/globe",package="threejs"))

# A Stand-alone example:
library("threejs")
library("maps")
data(world.cities, package="maps")
cities <- world.cities[order(world.cities$pop,decreasing=TRUE)[1:1000],]
value  <- 100 * cities$pop / max(cities$pop)

# Set up a color map, noting that THREE.Color only accepts RGB form
# so we need to drop the A value from the colors:
col <- sapply(heat.colors(10), function(x) substr(x,1,7))
names(col) <- c()
col <- col[floor(length(col)*(100-value)/100) + 1]

# Load the map of the world as a dataURI image using the \\code{texture}
# function. This is required for non-shiny use (shiny apps can just use
# the file name directly).
picture <- texture(system.file("htmlwidgets/lib/globe/world.jpg",package="threejs"))
globe.js(img=picture, lat=cities$lat, long=cities$long, value=value, color=col)

# Plot them on the moon
picture <- texture(system.file("images/moon.jpg",package="threejs"))
globe.js(img=picture, bodycolor="#555555", emissive="#444444",
         lightcolor="#555555", lat=cities$lat, long=cities$long,
         value=value, color=col)
}
\references{
The threejs project \url{http://threejs.org}.
The corresponding javascript file in
\code{ system.file("htmlwidgets/globe.js",package="threejs")}.

Includes ideas and images from the dat.globe Javascript WebGL Globe Toolkit
\url{http://dataarts.github.com/dat.globe},
Copyright 2011 Data Arts Team, Google Creative Lab
Licensed under the Apache License, Version 2.0
\url{http://www.apache.org/licenses/LICENSE-2.0}
}

