% Generated by roxygen2 (4.0.2): do not edit by hand
\name{helloThree.js}
\alias{helloThree.js}
\title{helloThree.js  Three.js hello world example.}
\usage{
helloThree.js(img = readPNG(system.file("img/Rlogo.png", package = "png")),
  repeatX = 8, repeatY = 8, ambient = "#999", specular = "#ffa",
  shininess = 10, height = NULL, width = NULL)
}
\arguments{
\item{img}{Either a matrix or raster image representation.}

\item{repeatX}{The number of times to repeat the image along the X axis.}

\item{repeatY}{The number of times to repeat the image along the Y axis.}

\item{ambient}{The intensity/color of object light scattering.}

\item{specular}{The intensity/color of object specular reflection.}

\item{shininess}{The shininess of the object.}

\item{height}{The container div height.}

\item{width}{The container div width.}
}
\description{
Three.js hello world example, maps an image onto a rotating sphere.
}
\examples{
## dontrun
# Stand-alone examples:
library("threejs")
helloThree.js(readPNG(system.file("img/Rlogo.png",package="png")))
helloThree.js(readJPEG(system.file("images/moon_1024.jpg",package="threejs")))

# A shiny example:
runApp(system.file("examples/hello",package="threejs"), launch.browser=FALSE)
}
\references{
The threejs project \url{http://threejs.org}.
The corresponding javascript file in
\code{ system.file("htmlwidgets/hello.js",package="threejs")}.
}

