% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-definitions.R
\docType{data}
\name{LeMis}
\alias{LeMis}
\title{Les Miserables Character Coappearance Data}
\format{
An igraph package graph object.
}
\source{
Mike Bostock's D3.js force-directed graph example  \url{https://bl.ocks.org/mbostock/4062045}.
Data based on character coappearence in Victor Hugo's Les Miserables, compiled by Donald Knuth
(\url{https://www-cs-faculty.stanford.edu/~uno/sgb.html}).
}
\usage{
data(LeMis)
}
\description{
Les Miserables Character Coappearance Data
}
\keyword{datasets}
