% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-definitions.R
\docType{data}
\name{pump}
\alias{pump}
\title{Pump rendering}
\format{A compressed JSON file.}
\source{
The threejs.org web site \url{https://threejs.org/examples/#webgl_animation_keyframes_json}
}
\description{
A keyframe animation example from the threejs.org website.
}
\keyword{animation}
\keyword{datasets}
\keyword{threejs}
