% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{threejs-package}
\alias{threejs}
\alias{threejs-package}
\title{Tools for creating threejs and d3.js graphics from R}
\description{
Tools for creating threejs and d3.js graphics from R
}
\examples{
## dontrun
library("shiny")
runApp(system.file("examples/globe",package="threejs"))
runApp(system.file("examples/scatterplot",package="threejs"))

# See also help for globe.js and scatterplot3.js
}
\references{
\url{http://threejs.org}
\url{http://www.html5rocks.com/en/tutorials/three/intro/}
}

