% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/threejs.R
\docType{package}
\name{threejs-package}
\alias{threejs}
\alias{threejs-package}
\title{Interactive 3D graphics including point clouds and globes using three.js and htmlwidgets.}
\description{
Interactive 3D graphics including point clouds and globes using three.js and htmlwidgets.
}
\examples{
\dontrun{
library("shiny")
runApp(system.file("examples/globe",package="threejs"))
runApp(system.file("examples/scatterplot",package="threejs"))

# See also help for globe.js and scatterplot3.js
}
}
\references{
\url{http://threejs.org}
\url{http://www.html5rocks.com/en/tutorials/three/intro/}
}

